pyproject.toml
[project]
name = "babycare-rag"
version = "0.1.0"
description = "A RAG-powered baby care assistant tool for team integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "faiss-cpu>=1.10.0",
    "google-genai>=1.9.0",
    "markitdown[all]>=0.1.1",
    "mcp[cli]>=1.6.0",
    "pillow>=11.1.0",
    "rich>=14.0.0",
    "scipy>=1.15.2",
    "tqdm>=4.67.1",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
babycare-rag-cli = "test_tools.cli_test:main"
babycare-rag-test = "test_tools.api_test:main"
